+++ Sockets/Include/sockets/io_error.h
@@ -6,17 +6,14 @@
 #define _IMPACT_IO_ERROR_H_
 
 #include <string>
-#include <exception>
+#include <stdexcept>
 
 namespace impact {
-  class io_error : public std::exception {
+  class io_error : public std::runtime_error {
   public:
     explicit io_error(const std::string& arg);
     virtual ~io_error() throw();
     virtual const char* what() const throw();
-
-  private:
-    std::string m_msg_;
   };
 }
