+++ Sockets/Source/basic_socket_async.cpp
@@ -1,0 +1,18 @@
+/**
+ * Created by TekuConcept on July 11, 2018
+ */
+
+#include "sockets/basic_socket.h"
+#include "basic_socket_common.inc"
+
+using namespace impact;
+
+// std::future<buffer_data>&
+// basic_socket::send_async(
+// 	const void*   __buffer,
+// 	int           __length,
+// 	message_flags __flags)
+// {
+// 	// static std::promise<async_buffer_data> p;
+// 	// return p.get_future();
+// }
