+++ Sockets/Source/io_error.cpp
@@ -7,10 +7,16 @@
 using namespace impact;
 
 io_error::io_error(const std::string& __arg)
-  : exception(), m_msg_(__arg) { }
+: std::runtime_error(__arg)
+{}
 
 
-io_error::~io_error() throw() { }
+io_error::~io_error() throw()
+{}
 
 
-const char* io_error::what() const throw() { return m_msg_.c_str(); }
+const char*
+io_error::what() const throw()
+{
+  return std::runtime_error::what();
+}
